test: updated test scripts into a single file
