+++ test/unit/register/testSignup.js
@@ -6,7 +6,7 @@ const jwt = require('jsonwebtoken');
 
 app.use(express.json());
 
-const port = process.env.PORT || 7000;
+const port = process.env.PORT;
 const server = app.listen(port, () =>
   console.log(`Listening on port ${port}...`)
 );
@@ -345,6 +345,292 @@ describe('Signup Fields Tests', () => {
     const confirmPassword = functions.confirmPassword('olaWale_17#');
     password.should.be.eq(confirmPassword);
 
+    done();
+  });
+});
+
+/**
+ *TEST 2
+ */
+
+// node assert module;
+var assert = require('assert');
+
+// mock confirm password function
+const confirmPassword = (str1, str2) => str1 === str2;
+
+//mock check string length function
+const checkLength = (str) => str.length > 5;
+
+// mock for email validation function
+const validateEmailField = (email) =>
+  /^(([^<>()\[\]\\.,;:\s@"]+(\.[^<>()\[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/.test(
+    email
+  );
+
+describe('Test the signup page', function () {
+  describe('Test to check if password mactch', () => {
+    it("should return false if password and confirmpassword don't match", () => {
+      assert.equal(confirmPassword('Hello', 'Hell'), false);
+    });
+    it("should return true if password and confirmpassword don't match", () => {
+      assert.equal(confirmPassword('Hello', 'Hello'), true);
+    });
+  });
+
+  describe('Test to check the length of string', () => {
+    it('should return false if length of string less than 5', () => {
+      assert.equal(checkLength('hell'), false);
+    });
+    it('should return true if length of string greater than 5', () => {
+      assert.equal(checkLength('Hellooo'), true);
+    });
+  });
+
+  describe('email should be a valid email', () => {
+    it('should return false if email is invalid', () => {
+      assert.equal(validateEmailField('hell@'), false);
+    });
+    it('should return true if email is valid', () => {
+      assert.equal(validateEmailField('hell00@test.com'), true);
+    });
+  });
+});
+
+/**
+ *
+ * TEST 3
+ */
+
+const request = require('supertest');
+
+let appTest;
+
+const userCredentials = {
+  first_name: 'Jane',
+  last_name: 'Doe',
+  email: 'johndoe@gmail.com',
+  phone: '07023455569',
+  password: 'garyTheSnail',
+  cpassword: 'garyTheSnail',
+};
+
+describe('POST /api/v1/registration', function () {
+  beforeEach(function (done) {
+    appTest = require('../../../server');
+    userCredentials;
+    done();
+  });
+
+  afterEach(async () => {
+    await appTest.close();
+  });
+
+  it('should verify if all fields are entered correctly', function (done) {
+    request
+      .agent(appTest)
+      .post('/register')
+      .send(userCredentials)
+      .expect(200)
+      .expect('Content-Type', /json/)
+      .expect('message', 'Registration successful');
+    done();
+  });
+
+  it('verify if it sends an error message if first name field is not filled and entered', function (done) {
+    request
+      .agent(appTest)
+      .post('/register')
+      .send({
+        last_name: 'Doe',
+        email: 'johndoe@gmail.com',
+        phone: '07023455569',
+        password: 'garyTheSnail',
+        cpassword: 'garyTheSnail',
+      })
+      .expect(400)
+      .expect('Content-Type', /json/)
+      .expect('msg', 'please include all fields');
+    done();
+  });
+
+  it('verify if it sends an error message if first name field contains only strings', function (done) {
+    request
+      .agent(appTest)
+      .post('/register')
+      .send({
+        first_name: '123Jane',
+        last_name: 'Doe',
+        email: 'johndoe@gmail.com',
+        phone: '07023455569',
+        password: 'garyTheSnail',
+        cpassword: 'garyTheSnail',
+      })
+      .expect(400)
+      .expect('Content-Type', /json/)
+      .expect('msg', 'please include all fields');
+    done();
+  });
+
+  it('verify if it sends an error message if last name field is not filled and entered', function (done) {
+    request
+      .agent(appTest)
+      .post('/register')
+      .send({
+        first_name: 'Jane',
+        // "last_name": "Doe",
+        email: 'johndoe@gmail.com',
+        phone: '07023455569',
+        password: 'garyTheSnail',
+        cpassword: 'garyTheSnail',
+      })
+      .expect(400)
+      .expect('Content-Type', /json/)
+      .expect('msg', 'please include all fields');
+    done();
+  });
+
+  it('verify if it sends an error message if last name field contains only strings', function (done) {
+    request
+      .agent(appTest)
+      .post('/register')
+      .send({
+        first_name: '123Jane',
+        last_name: 'D3546oe',
+        email: 'johndoe@gmail.com',
+        phone: '07023455569',
+        password: 'garyTheSnail',
+        cpassword: 'garyTheSnail',
+      })
+      .expect(400)
+      .expect('Content-Type', /json/)
+      .expect('msg', 'please include all fields');
+    done();
+  });
+
+  it('verify if it sends an error message if email field is not filled and entered', function (done) {
+    request
+      .agent(appTest)
+      .post('/register')
+      .send({
+        first_name: 'Jane',
+        last_name: 'Doe',
+        // "email": 'johndoe@gmail.com',
+        phone: '07023455569',
+        password: 'garyTheSnail',
+        cpassword: 'garyTheSnail',
+      })
+      .expect(400)
+      .expect('Content-Type', /json/)
+      .expect('msg', 'please include all fields');
+    done();
+  });
+
+  it('verify if it sends an error message if email field is not in the right format', function (done) {
+    request
+      .agent(appTest)
+      .post('/register')
+      .send({
+        first_name: 'Jane',
+        last_name: 'Doe',
+        email: 'johndoe',
+        phone: '07023455569',
+        password: 'garyTheSnail',
+        cpassword: 'garyTheSnail',
+      })
+      .expect(400)
+      .expect('Content-Type', /json/)
+      .expect('msg', 'please include all fields');
+    done();
+  });
+
+  it('verify if it sends an error message if phone field contains a string', function (done) {
+    request
+      .agent(appTest)
+      .post('/register')
+      .send({
+        first_name: 'Jane',
+        last_name: 'Doe',
+        email: 'johndoe',
+        phone: 'gary',
+        password: 'garyTheSnail',
+        cpassword: 'garyTheSnail',
+      })
+      .expect(400)
+      .expect('Content-Type', /json/)
+      .expect('msg', 'please include all fields');
+    done();
+  });
+
+  it('verify if it sends an error message if password field is empty', function (done) {
+    request
+      .agent(appTest)
+      .post('/register')
+      .send({
+        first_name: 'Jane',
+        last_name: 'Doe',
+        email: 'johndoe',
+        phone: '07023455569',
+        // 'password': 'garyTheSnail',
+        // "cpassword": 'garyTheSnail'
+      })
+      .expect(400)
+      .expect('Content-Type', /json/)
+      .expect('msg', 'please include all fields');
+    done();
+  });
+
+  it('verify if it sends an error message if password field does not math confirm password field', function (done) {
+    request
+      .agent(appTest)
+      .post('/register')
+      .send({
+        first_name: 'Jane',
+        last_name: 'Doe',
+        email: 'johndoe',
+        phone: '07023455569',
+        password: 'garyTheSnail',
+        cpassword: 'garyTheSn',
+      })
+      .expect(400)
+      .expect('Content-Type', /json/)
+      .expect('msg', "Password didn't correspond");
+    done();
+  });
+
+  it('verify if it sends an error message if password field is weak', function (done) {
+    request
+      .agent(appTest)
+      .post('/register')
+      .send({
+        first_name: 'Jane',
+        last_name: 'Doe',
+        email: 'johndoe',
+        phone: '07023455569',
+        password: 'ga',
+        cpassword: 'ga',
+      })
+      .expect(400)
+      .expect('Content-Type', /json/)
+      .expect('msg', 'Password is weak');
+    done();
+  });
+
+  it('verify if it sends an error message if all fields are empty', function (done) {
+    request
+      .agent(appTest)
+      .post('/register')
+      .send({
+        first_name: '',
+        last_name: '',
+        email: '',
+        phone: '',
+        password: '',
+        cpassword: '',
+      })
+      .expect(400)
+      .expect('Content-Type', /json/)
+      .expect('msg', 'please include all fields');
     done();
   });
 });
